refactor API handlers
