+++ server/routers/youtube.ts
@@ -1,31 +1,7 @@
 import * as trpc from '@trpc/server';
-import { google } from 'googleapis';
 import * as z from 'zod';
 import { publicProcedure, router } from '../trpc';
-
-async function getYouTubeChannelStatistics(channelId?: string[]) {
-  const auth = new google.auth.GoogleAuth({
-    credentials: {
-      client_email: process.env.GOOGLE_CLIENT_EMAIL,
-      private_key: process.env.GOOGLE_PRIVATE_KEY?.replace(/\\n/g, '\n') ?? '',
-    },
-    scopes: ['https://www.googleapis.com/auth/youtube.readonly'],
-  });
-
-  const youtube = google.youtube({
-    auth,
-    version: 'v3',
-  });
-
-  const response = await youtube.channels
-    .list({
-      id: channelId,
-      part: ['statistics'],
-    })
-    .then(res => res.data);
-
-  return response;
-}
+import { getYouTubeChannelStatistics } from '../data/youtube';
 
 export const youtubeRouter = router({
   getChannelStatistics: publicProcedure
